grid : moved Area related unit tests into their module
